nfc/nci: fix race with opening and closing
Previously we leverage NCI_UNREG and the lock inside nci_close_device to
prevent the race condition between  opening a device and closing a
device.

However, it still has problem because a failed  opening command will
erase the NCI_UNREG flag and allow another  opening command to bypass
the status checking.

This fix corrects that by making sure the NCI_UNREG is held.

Reported-by: syzbot+43475bf3cfbd6e41f5b7@syzkaller.appspotmail.com
Fixes: 48b71a9e66c2 ("NFC: add NCI_UNREG flag to eliminate the race")
Signed-off-by: f0rm2l1n <f0rm2l1n@foxmail.com>